feat: add structured data files for portfolio, activities, links, and resume
- Created new JSON files for activities, links, meta, portfolio, and resume data.
- Updated index.js to read and export new data files.
- Refactored index.html to dynamically render links, portfolio items, and activities from JSON data.
- Enhanced styling in style.scss for better presentation of activities and portfolio sections.
- Updated Vite configuration to include new data exports.
