+++ data/links.json
@@ -1,0 +1,22 @@
+[
+  {
+    "label": "ギットハブ",
+    "href": "https://github.com/nekobato"
+  },
+  {
+    "label": "エックス",
+    "href": "https://x.com/nekobato"
+  },
+  {
+    "label": "コセンス",
+    "href": "https://scrapbox.io/nekobatoken/"
+  },
+  {
+    "label": "ブログ",
+    "href": "https://sizu.me/nekobato/"
+  },
+  {
+    "label": "メール",
+    "href": "mailto:nekobato@gmail.com"
+  }
+]
