+++ data/meta.json
@@ -1,0 +1,8 @@
+{
+  "title": "nekobato.net",
+  "description": "nekobatoのホームページ",
+  "keywords": "nekobato.net",
+  "author": "nekobato",
+  "image": "https://nekobato.net/images/nekobato.png",
+  "url": "https://nekobato.net"
+}
