+++ data/resume-edu.json
@@ -1,0 +1,16 @@
+[
+  {
+    "institution": "慶応義塾大学",
+    "area": "環境情報学部",
+    "studyType": "学士",
+    "startDate": "2009-04",
+    "endDate": "2013-03"
+  },
+  {
+    "institution": "慶応義塾大学",
+    "area": "政策・メディア学科 CIプログラム",
+    "studyType": "修士",
+    "startDate": "2013-04",
+    "endDate": "2015-03"
+  }
+]
