+++ vite.config.js
@@ -1,6 +1,6 @@
 import { defineConfig } from "vite";
 import { ViteEjsPlugin } from "vite-plugin-ejs";
-import { meta, me, resume } from "./data.js";
+import { activities, links, meta, me, portfolio, resume } from "./data/index.js";
 
 export default defineConfig({
   plugins: [
@@ -8,6 +8,9 @@ export default defineConfig({
       meta,
       me,
       resume,
+      links,
+      portfolio,
+      activities,
     }),
   ],
 });
